nautilus_blockchain::rpc

Module utils

Functions

extract_rpc_subscription_id

Extracts the subscription ID from a blockchain RPC subscription event notification.

is_subscription_confirmation_response

Determines if a JSON message is a subscription response from the blockchain RPC server.

is_subscription_event

Determines if a JSON message is a subscription event notification from the blockchain RPC server.